requirements.txt
spotipy
streamlit
python-dotenv
google-generativeai
